+++ src/cuda_kernels/dim.h
@@ -11,12 +11,12 @@
 
 
 
-void mean_over_semilast_dim_backward(float *inp, float size, float *out,
+void mean_over_semilast_dim_backward(float *inp, int size, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node);
 
 
-void gather_last_dim_backward(float *inp, float dims_prod, float *out,
+void gather_last_dim_backward(float *inp, int dims_prod, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node);
