+++ src/cuda_kernels/tensor_scalar_codegen.h
@@ -8,6 +8,6 @@
 
 
 
-void scalarmult_backward(float *inp, float dims_prod, float *out,
+void scalarmult_backward(float *inp, int dims_prod, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node);
