+++ src/networks/modules/Conv2d/codegen.h
@@ -12,6 +12,6 @@
 #include "../../../mangler/scope_struct.h"
 
 
-void conv2d_backward(float *inp, float size, float *out,
+void conv2d_backward(float *inp, int size, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node);
