+++ src/networks/modules/Embedding/codegen.h
@@ -11,6 +11,6 @@
 
 
 
-void embedding_backward(float *inp, float size, float *out,
+void embedding_backward(float *inp, int size, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node);
