+++ src/networks/modules/LSTM/codegen.cpp
@@ -20,7 +20,7 @@
 
 
 
-void lstm_backward(float *inp, float size, float *out,
+void lstm_backward(float *inp, int size, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node)
 {
