+++ src/networks/modules/LSTM/codegen.h
@@ -10,6 +10,6 @@
 
 
 
-void lstm_backward(float *inp, float size, float *out,
+void lstm_backward(float *inp, int size, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node);
