+++ src/networks/modules/Linear/codegen.cpp
@@ -25,7 +25,7 @@
 
 
 
-void linear_backward(float *inp, float size, float *out,
+void linear_backward(float *inp, int size, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node)
 {
