+++ src/networks/modules/MaxPool2d/codegen.cpp
@@ -70,7 +70,7 @@ extern "C" void *Pool2d(Scope_Struct *scope_struct, DT_tensor *tensor)
 
 
 
-void pool2d_backward(float *inp, float size, float *out,
+void pool2d_backward(float *inp, int size, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node)
 {
