+++ src/networks/modules/MaxPool2d/codegen.h
@@ -10,6 +10,6 @@
 
 
 
-void pool2d_backward(float *inp, float size, float *out,
+void pool2d_backward(float *inp, int size, float *out,
                      float *dinp, float *dout,
                      std::string module_name, DT_tensor *node);
